refactor: 정렬 관련 유틸리티 클래스 패키지 변경
