+++ space-d/src/main/java/com/dnd/spaced/core/word/infrastructure/util/WordSortConditionConverter.java
@@ -2,7 +2,7 @@ package com.dnd.spaced.core.word.infrastructure.util;
 
 import static com.dnd.spaced.core.word.domain.QWord.word;
 
-import com.dnd.spaced.global.repository.OrderByNull;
+import com.dnd.spaced.global.persistence.OrderByNull;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.ComparableExpressionBase;
 import java.util.ArrayList;
