+++ space-d/src/main/java/com/dnd/spaced/global/persistence/OrderByNull.java
@@ -1,11 +1,11 @@
-package com.dnd.spaced.global.repository;
+package com.dnd.spaced.global.persistence;
 
 import com.querydsl.core.types.NullExpression;
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
 
 @SuppressWarnings({"rawtypes", "unchecked"})
-public class OrderByNull extends OrderSpecifier {
+public final class OrderByNull extends OrderSpecifier {
 
     public static OrderByNull DEFAULT = new OrderByNull();
